requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv
pandas